[Close #1] Refine code format
Add code check scripts for pull request.
