+++ requirements.txt
@@ -1,0 +1,5 @@
+pandas
+plotly
+requests
+scipy
+numpy
