+++ tox.ini
@@ -1,0 +1,13 @@
+[tox]
+minversion=2.3.1
+envlist = py37, black
+skipsdist=True
+skip_missing_interpreters=True
+
+[testenv:black]
+deps=black
+basepython=python3
+setenv =
+    LC_ALL=C.UTF-8
+    LANG=C.UTF-8
+commands=black --check --verbose -l 100 src
